Allows for infinite scrolling of slides.
This adds a flag to the options that when set to true will repeat the
slides infinitely. This means that when scrolling if you are at the last
slide and scroll for the next one it will display the first slide as if
it was next in line. Similarly if you are on the first slide and scroll
for the prev slide it will display the last as if it was next in line.
Sadly, this adds a jquery dependency as it was the cleanest way to
manipulate the DOM to add the duplicate data required to allow for the
infinite scrolling to be seamless.