Add 'preventClick' option
In certain use cases (such as with jQuery Mobile and a fixed header), it
is desirable to prevent the click event from propagating up through the
document. This options provides a way to configure that behavior; by
default, it is off and will click events to propagate per normal
behavior.