Fix argument error in <= IE9
Was getting an argument error in <= IE9 until I change the call a bit to check for a function first.  Looked like fn was actually 'true' instead of a function on first call...  Weirdly enough I am not getting the same error on the swipejs home page but the slideshow doesn't actually finish going through all the slides, possibly another issue with <= IE9?